updated the advert migration

+++ db/migrate/20200922163823_create_adverts.rb
@@ -2,7 +2,7 @@ class CreateAdverts < ActiveRecord::Migration[6.0]
   def change
     create_table :adverts do |t|
       t.string :url
-      t.integer :source_id
+      t.bigint :source_id
       t.string :source_name
       t.references :car, null: false, foreign_key: true
